crawler/pyproject.toml
[tool.poetry]
name = "danakka-crawler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "danakka_crawler"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.2"
chromedriver-autoinstaller = "^0.4.0"
undetected-chromedriver = "3.2.1"
webdriver-manager = "^3.8.5"
lxml = "^4.9.2"
six = "^1.16.0"
cssutils = "^2.6.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
